feat: sponsorship auth

+++ backend/src/controllers/sponsorshipController.js
@@ -4,35 +4,36 @@ import { configurationModel } from '../models/configurationModel.js';
 import { RecordNotFound } from '../shared/exceptions.js';
 import { BasicState, Roles } from '../shared/enums.js';
 
-export const findAllSponsorships = async (req, res, next) => {
+export const findAllSponsorships = async (req, res) => {
   try {
-    const sponsorships = await sponsorshipModel.find({});
-    res.json(sponsorships);
-  } catch (error) {
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
-  }
-};
+    const { actor } = res.locals;
 
-export const findSponsorship = async (req, res, next) => {
-  try {
-    const sponsorship = await sponsorshipModel.findById(req.params.sponsorshipId);
-
-    if (!sponsorship) {
-      return next(new RecordNotFound());
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized');
     }
 
-    res.json(sponsorship);
+    if (actor.role !== Roles.ADMIN) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation');
+    }
+
+    const sponsorships = await sponsorshipModel.find({});
+    return res.json(sponsorships);
   } catch (error) {
     res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
   }
 };
 
 export const createSponsorship = async (req, res) => {
-  const newSponsorship = new sponsorshipModel(req.body);
-
   try {
+    const { actor } = res.locals;
+    
+    if (!(actor.role === Roles.SPONSOR || actor.role === Roles.ADMIN)) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation');
+    }
+    
+    const newSponsorship = new sponsorshipModel(req.body);
     const sponsorship = await newSponsorship.save();
-    res.status(StatusCodes.CREATED).json(sponsorship);
+    return res.status(StatusCodes.CREATED).json(sponsorship);
   } catch (error) {
     if (error.name === 'ValidationError') {
       res.status(StatusCodes.UNPROCESSABLE_ENTITY).json(error);
@@ -42,12 +43,66 @@ export const createSponsorship = async (req, res) => {
   }
 };
 
+export const findMySponsorships = async (req, res) => {
+  try {
+    const { actor } = res.locals;
+
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized');
+    }
+
+    if (!(actor.role === Roles.SPONSOR || actor.role === Roles.ADMIN)) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation');
+    }
+
+    const sponsorships = await sponsorshipModel.find({ sponsor: actor._id });
+    return res.json(sponsorships);
+  } catch (error) {
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
+  }
+};
+
+export const findSponsorship = async (req, res, next) => {
+  try {
+    const sponsorship = await sponsorshipModel.findById(req.params.sponsorshipId);
+    const { actor } = res.locals;
+
+    if (!sponsorship) {
+      return next(new RecordNotFound());
+    }
+
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized');
+    }
+
+    if (!(actor.role === Roles.ADMIN || actor._id.toString() === sponsorship.sponsor.toString())) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation.');
+    }
+
+    return res.json(sponsorship);
+  } catch (error) {
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
+  }
+};
+
+
 export const updateSponsorship = async (req, res) => {
   try {
+    const { sponsor } = await sponsorshipModel.findById(req.params.sponsorshipId)
+    const { actor } = res.locals;
+
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized.');
+    }
+
+    if (!(actor.role === Roles.ADMIN || actor._id.toString() === sponsor.toString())) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation.');
+    }
+
     const sponsorship = await sponsorshipModel.findOneAndUpdate({ _id: req.params.sponsorshipId }, req.body, {
       new: true
     });
-    res.json(sponsorship);
+    return res.json(sponsorship);
   } catch (error) {
     if (error.name === 'ValidationError') {
       res.status(StatusCodes.UNPROCESSABLE_ENTITY).json(error);
@@ -59,8 +114,19 @@ export const updateSponsorship = async (req, res) => {
 
 export const deleteSponsorship = async (req, res) => {
   try {
+    const { sponsor } = await sponsorshipModel.findById(req.params.sponsorshipId)
+    const { actor } = res.locals;
+
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized.');
+    }
+
+    if (!(actor.role === Roles.ADMIN || actor._id.toString() === sponsor.toString())) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation.');
+    }
+
     await sponsorshipModel.deleteOne({ _id: req.params.sponsorshipId });
-    res.sendStatus(StatusCodes.NO_CONTENT);
+    return res.sendStatus(StatusCodes.NO_CONTENT);
   } catch (error) {
     res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
   }
@@ -68,17 +134,27 @@ export const deleteSponsorship = async (req, res) => {
 
 export const paySponsorship = async (req, res) => {
   try {
+    const { sponsor } = await sponsorshipModel.findById(req.params.sponsorshipId)
+    const { actor } = res.locals;
     const isPaymentApproved = true; // Payment logic and connection with Paypal
+
+    if (!actor) {
+      return res.status(StatusCodes.UNAUTHORIZED).send('Not authorized.');
+    }
+
+    if (!(actor.role === Roles.ADMIN || actor._id.toString() === sponsor.toString())) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).send('You cannot perform this operation.');
+    }
 
     if (isPaymentApproved) {
       const activeSponsorship = await sponsorshipModel.findOneAndUpdate(
         { _id: req.params.sponsorshipId },
         { state: BasicState.ACTIVE }
       );
-      res.json(activeSponsorship);
-    } else {
-      res.status(StatusCodes.SERVICE_UNAVAILABLE).send({ message: 'Error processing payment.' });
+      return res.json(activeSponsorship);
     }
+      
+    return res.status(StatusCodes.SERVICE_UNAVAILABLE).send({ message: 'Error processing payment.' });
   } catch (error) {
     res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
   }
